requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
amply==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
chempy==0.7.12
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.4.1
docutils==0.16
dot2tex==2.11.3
entrypoints==0.3
futures==3.1.1
googletrans==3.0.0
goslate==1.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.9
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
more-itertools==8.5.0
mpmath==1.1.0
multidict==4.7.6
nbclient==0.5.0
nbconvert==6.0.1
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.1
numpydoc==1.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
PuLP==2.3
py==1.9.0
pycparser==2.20
PyDictionary==2.0.1
pyflakes==2.2.0
Pygments==2.6.1
pyneqsys==0.5.6
pyodesys==0.13.1
pyparsing==2.4.7
pyrsistent==0.17.2
pytest==6.0.1
pytest-flakes==4.0.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
quantities==0.12.4
requests==2.24.0
rfc3986==1.4.0
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
sniffio==1.1.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sym==0.3.4
sympy==1.6.2
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wikipedia==1.4.0
yarl==1.5.1
zipp==3.1.0
